cleaning up the garbage fire
