+++ constants/Icons.js
@@ -1,0 +1,13 @@
+const EMOJIS = {
+    '-none':'',
+    'bread': '🥖',
+    'eggs': '🥚',
+    'milk': '🥛', 
+    'prepared': '🥡',
+    'produce': '🥦',
+    'shelf': '🥫',
+  }
+
+  export default {
+    EMOJIS
+  }
